LegalEaseAI/requirements.txt
streamlit
PyPDF2
python-docx
ai71
openai
langchain
langchain_community
langchain_openai
